Update docs for tpm2 and tpm2_client crates.
- Add documentation where missing
- Format to wrap to 80 characters
- Move some tpm2_client code into a protocol module to clean up
  the organization of the docs.
