+++ crates/tpm2-client/src/connection/mod.rs
@@ -1,13 +1,11 @@
-//! This module provides traits to communicate with a
-//! TPM via a particular medium. The top-level trait
-//! is [`Connection`].
+//! The connection module provides the top level [`Connection`] trait and
+//! common implementations (as submodules) of the trait.
 
 use core::error::Error;
 
-mod tcp;
-pub use tcp::*;
+pub mod tcp;
 
-/// Trait for communicating with a TPM.
+/// Common trait for communicating with a TPM.
 pub trait Connection {
     /// The type returned if [`Connection::transact`] fails.
     ///
@@ -15,9 +13,11 @@ pub trait Connection {
     /// connection itself. If the connection can never fail, this can be
     /// [`Infallible`](core::convert::Infallible).
     type Error: Error;
+
     /// Perform a command/response transaction with the TPM.
     ///
-    /// Returns a slice of the response containing the bytes that were written.
+    /// Returns a slice of the response containing the bytes that were returned
+    /// from the TPM.
     ///
     /// Note that even if the response contains a `TPM_RC` error, this method
     /// still returns `Ok(...)`. `Err` is only returned when we are unable to
