+++ crates/tpm2-client/tests/simulator.rs
@@ -11,7 +11,7 @@ use std::io::Result;
 use std::net::Ipv4Addr;
 
 use tempfile::tempdir;
-use tpm2_client::connection::TcpSimulator;
+use tpm2_client::connection::tcp::TcpSimulator;
 use tpm2_client::run_command;
 use tpm2_rs_base::commands::StartupCmd;
 use tpm2_rs_base::constants::TpmSu;
